LayoutTests/imported/w3c/web-platform-tests/screen-wake-lock/wakelock-active-document.https.window-expected.txt


PASS navigator.wakeLock.request() aborts if the document becomes not active.
PASS navigator.wakeLock.request() aborts if the document is not fully active.
PASS navigator.wakeLock.request() aborts if the document is active, but not fully active.
